Move write-only helper functions and tests to their own files.
